requirements.txt
mcp[cli]>=1.9.0
google-auth>=2.29.0
google-api-python-client>=2.130.0
pydantic>=2.7.0
